requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
pytest==7.3.0